+++ R/med_mod_plots.R
@@ -15,8 +15,8 @@
 #' Plot mediation model with one or more mediators
 #'
 #' Returns graphViz code and graph for (multiple) mediation model.
-#' Coefficients and significance  values for paths need to be provided -
-#' see example for the format.
+#' Coefficients and significance values for paths need to be provided, 
+#' based on the format returned by `run_moderation()`
 #'
 #' @encoding UTF-8
 #' @param IV Character. Name of predictor
@@ -27,7 +27,7 @@
 #' provided, function will align mediators automatically, which is unlikely to
 #' provide a well-aligned path (except for cases when offset has been implemented
 #' for that number of mediators, currently 1 and 3). However, returned code can still
-#' be edited. See rNuggets:::.coef_offset_3 for an example of an offset tibble.
+#' be edited. See timesaveR:::.coef_offset_3 for an example of an offset tibble.
 #' @param digits Number of digits for rounding
 #' @param filename If provided, graph will be saved as .svg file.
 #' @param ind_p_values Should significance stars be shown for indirect effects,
@@ -37,20 +37,14 @@
 #' @export
 #' @examples
 #' \dontrun{
-#' # Values for model
-#' med_model <- tibble::tribble(
-#'   ~type, ~mediator, ~est, ~pvalue, ~ci.lower, ~ci.upper,
-#'   "a", "Empathy", 0.29, 0, 0.29, 0.29,
-#'   "b", "Empathy", 0.18, 0, 0.18, 0.18,
-#'   "direct", NA, -0.08, 0, -0.08, -0.08,
-#'   "indirect", "Empathy", 0.05, 0, 0.05, 0.05,
-#'   "total", NA, 0.41, 0, 0.41, 0.41
-#' )
+#' # Estimate mediation model
 #'
+#' res <- run_mediation(ess_health, fltdpr, health, dosprt, agea, bootstraps = 100)
+#' 
 #' # Run plot command
-#' rNuggets:::plot_mediation(
-#'   IV = "Positive <br /> contact",
-#'   DV = "Diversity <br /> policies", Ms = "Empathy", df = med_model
+#' plot_mediation(
+#'   IV = "Frequency of  <br /> feeling depressed",
+#'   DV = "Self-reported <br /> poor health", Ms = "Frequency of <br /> physical activity", df = res
 #' )
 #' }
 plot_mediation <- function(IV, DV, Ms, df, digits = 2, coef_offset = length(Ms), filename = NULL, ind_p_values = FALSE) {
@@ -202,7 +196,7 @@ plot_mediation <- function(IV, DV, Ms, df, digits = 2, coef_offset = length(Ms),
 }
 
 
-moderated_mediation <- function(X, M, W, Y, CV = NULL, mod_direct_path = TRUE, labels = list(a = "+", b = "+", c = "+", a_mod = "+", c_mod = "+"), filename = NULL) {
+plot_moderated_mediation <- function(X, M, W, Y, CV = NULL, mod_direct_path = TRUE, labels = list(a = "+", b = "+", c = "+", a_mod = "+", c_mod = "+"), filename = NULL) {
   .check_req_packages(c("glue", "DiagrammeR"))
 
   all_text <- paste(X, M, W, Y, CV)